// client/react/theme/ITheme.tsx
export interface ITheme {
    primary: {
        Normal: string;
        LightVariant: string;
        DarkVariant: string;
        On: string;
    };
    secondary: {
        Normal: string;
        LightVariant: string;
        DarkVariant: string;
        On: string;
    };
    info: {
        Normal: string;
        LightVariant: string;
        DarkVariant: string;
        On: string;
    };
    success: {
        Normal: string;
        LightVariant: string;
        DarkVariant: string;
        On: string;
    };
    warning: {
        Normal: string;
        LightVariant: string;
        DarkVariant: string;
        On: string;
    };
    error: {
        Normal: string;
        LightVariant: string;
        DarkVariant: string;
        On: string;
    };
    background: {
        Normal: string;
        On: string;
    };
    surface: {
        Normal: string;
        On: string;
    };
    divider: {
        Normal: string; // 分割线
        Highlight: string; // 高亮分割线
    }
}